Fixed ClassCastException

+++ project/src/main/scala/pcd/assignment03/distributed_programming/model/Point2D.scala
@@ -7,18 +7,18 @@ trait Point2D:
   /**
    * X of the Point2D
    */
-  def x: Double
+  def x: Int
 
   /**
    * Y of the Point2D
    */
-  def y: Double
+  def y: Int
 
 object Point2D:
-  def apply(x: Double, y: Double): Point2D = Point2DImpl(x, y)
+  def apply(x: Int, y: Int): Point2D = Point2DImpl(x, y)
 
   private case class Point2DImpl(
-                                  override val x: Double,
-                                  override val y: Double
+                                  override val x: Int,
+                                  override val y: Int
                                 ) extends Point2D
